+++ src/omneo/index.ts
@@ -8,6 +8,7 @@ import Locations from './resources/locations/index.js'
 import Auth from './resources/auth/index.js'
 import Tenants from './resources/tenants/index.js'
 import Connections from './resources/connections'
+import Identities from './resources/identities'
 import Interactions from './resources/interactions'
 
 export class Omneo {
@@ -35,6 +36,7 @@ export class Omneo {
   public tenants = new Tenants(this)
   public connections = new Connections(this)
   public interactions = new Interactions(this)
+  public identities = new Identities(this)
 
   async call (requestParams: OmneoRequest): Promise<any> {
     const { endpoint, params = {}, method, body } = requestParams